+++ salesforce/package.json
@@ -17,6 +17,8 @@
     "next": "12.0.9",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-sortablejs": "^6.1.0",
+    "sortablejs": "^1.14.0",
     "tailwindcss": "^3.0.17"
   },
   "devDependencies": {
@@ -25,6 +27,7 @@
     "@types/enzyme": "^3.10.11",
     "@types/node": "17.0.13",
     "@types/react": "17.0.38",
+    "@types/sortablejs": "^1.10.7",
     "autoprefixer": "^10.4.2",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
