+++ salesforce/src/styles/globals.css
@@ -1,3 +1,7 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
+
+.sortable-chosen {
+  box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
+}
